mainboard/supermicro/h8scm: Mainboard added
Base files were imported from coreboot 4.7 and updated to work with
coreboot 4.11. Support for native fan control by the hwmon was added.
The board works well and boots in about 8 seconds with 1x32GB RDIMM.

There is still no serial output, the CBMEM console was used to debug
and configure the hwmon correctly.